+++ src/components/LoadMore.vue
@@ -36,7 +36,7 @@
                 }
                 this.searchStore.stop()
 
-                this.value++
+                this.value = this.current + 1
                 this.searchStore.loadMore()
 
                 this.searchStore.start()